requirements.txt
alembic==1.0.11
autopep8==1.4.4
baidu-aip==2.2.17.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.13
MarkupSafe==1.1.1
mysqlclient==1.4.2
pycodestyle==2.5.0
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.5
strutil==0.2.1
urllib3==1.25.3
Werkzeug==0.15.4
